Adaptation of prepare_data.py and generate_tilesets.py
Adaptation of prepare_data.py and generate_tilesets.py from a mono-class to a multi-class case.
Added the creation of the attributes CATEGORY and SUPERCATEGORY to the swimmingpools dataset.
Earesed the category info from config_GE.yaml
Added the category determination to the function get_COCO_image_and_segmentations()
Modified some code for the management of multiple categories lines 455 to 505 (new lines count)