1st try completed multi
First try at a complete multi-class detector. It runs correctly for the multi- and mono-class examples and should give all the necessary indicators. However, the FN and recall results are intriguing. In addition, the classes in the end are identified by a contiguous indicator, which is not the easiest to understand for users.