Correct the score threshold setting position in script
The position of the score threshold setting in the script was not correct. The threshold score set in the config file was not taken into account. Instead the value set in detectron2 config file was taken into account. The .pkl file name was correct with the score value set in the config file taken into account. The attribution of the threshold score must occur before the definition of the predictor.


Former-commit-id: 6c2eea3aa64dfd11ad6f2a1904553e8a77ad75ae